ghala/requirements.txt
asgiref==3.7.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
packaging==23.1
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
